+++ src/Util/ZenViewUtil.ts
@@ -1,6 +1,10 @@
 import * as vscode from 'vscode';
-import { ZenViewFileBuilder } from './ZenViewFileBuilder';
+import { ZenViewFileBuilder } from '../ZenViewFileBuilder';
 import * as Path from 'path';
+import { ZenViewQuickPickItem } from '../ZenViewQuickPickItem';
+import { ZenViewFile } from '../ZenViewFile';
+import { zenViewGlobals } from '../ZenViewGlobals';
+import { ZenFileSystemHandler } from '../ZenFileSystemHandler';
 
 export class ZenViewUtil {
 
@@ -62,6 +66,30 @@ export class ZenViewUtil {
             }
         }
         return fileType;
+    }
+
+    public convertZenViewFilesToQuickPickItems(files: ZenViewFile[]): ZenViewQuickPickItem[] {
+        let items: ZenViewQuickPickItem[] = [];
+        for (let file of files) {
+            items.push(new ZenViewQuickPickItem(file.fileName, file.fileUri, file.fileType));
+        }
+        return items;
+    }
+
+    public async getAllZenFiles(): Promise<ZenViewFile[]> {
+        let zenFiles: ZenViewFile[] = [];
+        for (let path of zenViewGlobals.zenPaths) {
+            if (path.fileType === vscode.FileType.Directory) {
+                let files = ZenFileSystemHandler.getFilesRecursive(vscode.Uri.file(path.fileUri));
+                for await (let file of files) {
+                    zenFiles.push(file);
+                }
+            }
+            else {
+                zenFiles.push(path);
+            }
+        }
+        return zenFiles;
     }
 }