+++ src/ZenViewGlobals.ts
@@ -1,12 +1,14 @@
 import { ConfigHandler } from "./ConfigHandler";
 import * as vscode from 'vscode';
 import { ZenViewFile } from "./ZenViewFile";
+import { SearchResult } from "./SearchAlgorithm";
 
 export class ZenViewGlobals {
     public rootPath: vscode.Uri | undefined;
-    public zenPaths! : ZenViewFile[];
+    public zenPaths!: ZenViewFile[];
+    public searchResults: SearchResult[] = [];
 
-    constructor() {    
+    constructor() {
         this.rootPath = vscode.workspace.workspaceFolders![0].uri;
         this.onConfigChange();
     }